requirements.txt
numpy
matplotlib
requests
python-twitter
imageio
